fix(Sequelize): :bug: totalPrice and price changed for decimal type than float

+++ back-end/src/database/migrations/20220920192726-sales.js
@@ -33,7 +33,7 @@ module.exports = {
       },
       totalPrice: {
         allowNull: false,
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(9, 2),
         field: 'total_price',
       },
       deliveryAdress: {
